+++ src/parchment_intake/models.py
@@ -9,10 +9,10 @@ class ParchmentIntake(models.Model):
     supervisor = models.ForeignKey(Employee, on_delete=models.SET_NULL, null=True)
     supplier = models.ForeignKey(Supplier, on_delete=models.SET_NULL, null=True)
     
-    total_bags_count = models.IntegerField()
+    total_bags_count = models.FloatField()
     total_weight = models.FloatField()
     moisture_content = models.FloatField(default=0.00)    
-    sample_result = models.FloatField(default=0)    
+    sample_result = models.FloatField(default=0.00)    
 
     resting_period = models.IntegerField()
